refactor: use generated Numeric type

+++ adder/src/lib.rs
@@ -1,11 +1,12 @@
 mod bindings;
 
+use crate::bindings::calc::basic::types::Numeric;
 use crate::bindings::exports::calc::basic::add::Guest;
 
 struct Component;
 
 impl Guest for Component {
-    fn add(a: f64, b: f64) -> f64 {
+    fn add(a: Numeric, b: Numeric) -> Numeric {
         a + b
     }
 }
