+++ app/src/main.rs
@@ -1,5 +1,6 @@
 mod bindings;
 
+use crate::bindings::calc::basic::types::Numeric;
 use crate::bindings::calc::basic::calculate::{eval_expression, Op};
 
 fn main() {
@@ -8,8 +9,8 @@ fn main() {
     let _ = args.next(); // ignore first arg as it is the program name
 
     // TODO: handle erroneous inputs
-    let x = args.next().unwrap().parse::<f64>().unwrap();
-    let y = args.next().unwrap().parse::<f64>().unwrap();
+    let x = args.next().unwrap().parse::<Numeric>().unwrap();
+    let y = args.next().unwrap().parse::<Numeric>().unwrap();
     let op = args.next().unwrap();
 
     let result = match op.as_str() {
